That renaming is surprisingly easy.
