+++ src/Sound/Dummy/Environment.cpp
