+++ src/Sound/Dummy/Environment.h
